Работа с GUID вынесена в BinaryGuid.
