+++ src/tool1cd/Common.h
@@ -8,12 +8,6 @@
 
 void time1CD_to_FileTime(System::FILETIME *ft, const void *time1CD);
 unsigned int reverse_byte_order(unsigned int value);
-String GUIDas1C(const unsigned char* fr);
-String GUIDasMS(const unsigned char* fr);
-String GUID_to_string(const System::TGUID& guid);
-bool string_to_GUID(const String& str, System::TGUID* guid);
-String GUID_to_string_flat(System::TGUID* guid);
-bool string_to_GUID_flat(const String& str, System::TGUID* guid);
 bool two_hex_digits_to_byte(const wchar_t hi, const wchar_t lo, unsigned char& res);
 bool string1C_to_date(const String &str, void *bytedate);
 bool string_to_date(const String &str, void *bytedate);
